EMSE6992 final project
This is the original code for final project under python 3. This code describes a marathon record with both previous half and the final grade. I performed analysis on it to show possible reasons relate to the final result. To run this properly, you need to download the csv file called 'marathondata.csv' and find the absolute route in your own environment.